refactor: converting to native modules
